Improvements (#3)
negspacy is now locally available (as long as my changes are not integrated into the upstream repo). during document processing and chunk set building noun_chunks will be tagged as negated or not and when the connection between chunk text and documents is tracked, documents will be omitted for negated chunks.